Decouple and refactor back-end API to libtmux
