+++ tests/test_session.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, print_function,
 import logging
 import pytest
 
-from tmuxp import Pane, Session, Window
+from libtmux import Pane, Session, Window
 
 from .helpers import TEST_SESSION_PREFIX, namer
